fix(radio): move the sort picker beside the header actions
The sort picker sat in the toolbar below the header, on its own line under the browse and add buttons. It now shares the header row with them, and the toolbar keeps only the filter chips.
